// src/services/PriceItemService/models/PriceItem.ts
export interface PriceItemDto {
  id: number;
  date: string;
  cost: number;
  costInDollar: number;
  costInEuro: number;
  margin: number;
  publicDate: string;
  interName: string;
  groupName: string;
  manufacturer: string;
  country: string;
  distributor: string;
  priceId: number;
  productId: number;
  name: string;
  distributorId: number;
}